[TASK] Don't disable date selectbox on date-first
When picking a date from the agenda (date-first), the selectbox is
disabled. While explained in the CSH that this is because the date was
picked from the agenda instead of the selectbox, this is very
counter-intuitive. Also, this was the solution chosen because you can
pick a date very far into the future, but this messes up all kinds of
stuff with the conditions, so we'll have to consider that a date that
isn't allowed for any type shouldn't be selectable in the first place,
in a later commit.

Change-Id: I0000000000000000000000000000000000000000