[BUGFIX] Error if appointment property is deleted
When creating or editing an appointment, the form is built with uids
pointing towards the address and formfieldvalue property objects. If
these are removed somehow (e.g. TCA) before the feuser submits the form
to createAction or updateAction, PHP produces a Fatal Error because
validation fails on non-object properties and THEN attempts to call
functions from them! It would be better if there was a nice redirect and
explanation.

Change-Id: Ib79c410487461dce612f55035fd605f35b57ef54