[BUGFIX] Changed type->formfields does not always show them in FE
If an appointment was saved with 5 formfieldvalues, and these formfields
were removed from the type later on, to be replaced with 5 new ones, you
wouldn't see any formfield on the edit form. You could NEVER add the new
formfieldvalues while this is true. This problem only occurs when the
amount of formfields stays the same after a change.

Change-Id: I43c5d5fed7a4501d1688cc349113c5f3effe482f