[TASK] Make js GET request more efficient
The GET request is a FULL GET request, while we also have HEAD as
option, which gets exactly what we need: the headers. Also, the get
request should have asynchronous, because it causes a delay right now,
while setting it asynchronous will let it be done roughly around the
same time as the entire page load, which is good enough!

Change-Id: I2d92f4802ece995ec096d632d092356d9c7dd1e7