[TASK] remove back-link session clear
Cleaning the session on clicking the back link, will clean it also when
staying on the page after a warn-unload message. Might as well remove
it, since new and edit actions already clear the session. This way, we
open up the possibility to create some sort of undo while the session
lasts :)

Change-Id: I83e73a017bd135c3f14cc4ef8fbd281ca83384ac