[TASK] Set properties lazy when appropriate
A lot of properties (especially appointments') aren't set lazy because I
didn't know I could load the real instances so easily. The truth is that
especially on appointments, there are as much use cases where properties
aren't requested as where they are, so there is a serious
performance-benefit to be gained.

Change-Id: Ib435d7252f8c3ed5bbc285ac728d2450ea357086