[TASK] Restructure Controllers
Do some decoupling on certain overrides set in
SettingsOverrideController that aren't specific to "settings override",
and chain inherit them all. Also, make it so that there is an
appointments specific controller that has all the appointments-specific
functions, including functions that are currently only in appointment,
but should also be in agenda controller, then extend that one in both
the domain controllers.

Change-Id: I5163c20f74d7397ffeed266d70fc525e7d85248d