[TASK] Enable/disable for address fields
TCA, sql, email-building and template changes required, but also some
changes to validation. Solving it with the variableValidator is not an
option as that brings problems because appointments is a method argument
and not a domain property. Create a domain validator instead that
specifically sets the validator based on TCA setting.

Change-Id: Ia0bdd001890020f524ffade5431a97195fc3042b