[BUGFIX] Hiding/deleting a type results in exceptions
Hiding or deleting a type that is in use by appointments which show on
the listAction, will result in errors because the templates try to
interact with the then empty type. If a type does not exist to the
frontend, it simply should not show the appointments that belong to it.

Change-Id: I4a33aa1edece20616f865ad0a6b1be58f9c9629c