[BUGFIX] firstAvailableTime overridden when an appointment exists
When an appointment exists, slotService calculates time around the
blocks of time that are taken by said appointment. It then sets the
timestamp to the end of the block, and starts counting again until the
beginning of the next block. Timestamp is initially set to
firstAvailableTime, and uses that as startingpoint. However, if a
block-end is lower than firstAvailableTime, it does not differentiate
and still sets timestamp to it, overriding its initial value no matter
what it was. This should be fixed by checking each block-end when
defining the blocks, if they are smaller. If so, the block-end should
receive the timestamp value instead.

Change-Id: Idc96aa426606da9df4c90dc3d65b73fd8db7107b