[TASK] Allow different types of timeslot shifting
Available timeslots are always shifted based on appointment-type
intervals, but this often results in the firstavailabletime to be
somewhere mid weekday and then suddenly not available because its
interval results in a shift. This is annoying, especially when the
interval is small, so there should at least be a method that will shift
only at midnight or equivalent.

To pass that setting to the slotService in a similar way as
expireMinutes however, would severely increase complexity, because not
the service, but the controller has to read settings. So at the same
time, I'd like it if slotService gets an intialize method which passes
said settings instead. This would also require quite some cleanup due to
expireMinutes.

Change-Id: Ia010c673811c465824462b82d64f04db66eda6bd