[TASK] Less persisting, more intelligent resetting
EVERY add/update/remove is followed by a resetObjectStorage() call, in
which a persistChanges() call resides. Although the latter is already
better than the previous persistAll(), it still happens more than
necessary since the newAction() was split in multiple actions.
Currently, it is only needed @ add() due to object serialization @
redirect(), and @ clearExpiredAppointments() due to rebuilding
slotStorages afterwards.

Complementing persistence changes, reset can then be placed WITHIN the
repository functions so that most of these calls are automatic and
defined less redundantly. The only issue with this, is that
clearExpiredAppointments() does an update() within a loop, which could
result in multiple resets of a single type and thus redundant cache
queries. Giving a boolean switch to update() which decides whether to
reset, seems adequate. An if is less costly than a query.

Change-Id: I5e6177f06b75e718066aed8319cd721de0ecbbd0