[TASK] Assign types through agenda instead of plugin
Setting types through plugin never really "associates" types with
agenda's, as another plugin elsewhere might have the same agenda with
different types. This poses a problem @ caching, as reserving a timeslot
needs to influence ALL appointment types that are associated with the
agenda. Could not possibly be correct until this patch.

Change-Id: I54f8f61679d3570774bb623728ffe6fc47dea47c