[TASK] Storage object should be reset for all agenda types
Now that types are assigned to agenda instead of plugin, we can reset
all types when a storageobject needs to be reset. While this now works,
THIS PATCH IS NOT OPTIMAL AS IT ALSO RESETS EXCLUSIVE AVAILABILITY
TYPES. A new task is created to solve that issue, so that the current
patch can already be tested.

Change-Id: Ib47c66617f89c7b552a67b4e6c837fe188fa46d1