[TASK] Mark addtime fields as well @ failtime
We have a nice function now, all we need to do is add them to it like
the validator does. The only problem is that there seems to be a bug in
my way of handling the appointment validation errors, that never
appeared because validation errors are always within a single property:
formfieldvalues. You see, having TWO appointment properties, will result
in two times the ul because of the for attribute of the form.errors VH.
So that would also require a small workaround.

Change-Id: I9941e09490fe12962895ede40662e0d3c3780ce1