[BUGFIX] File upload form 'formObjectName' errors @ TYPO3 6.x
All fields are replaced by the same error message when
rewrittenPropertyMapper is turned off. Turns out that the
ViewHelperContainer can't get the form class by its original
non-namespaced name, and appointments is affected by this because it
uses custom form ViewHelpers which retain the original class-names. A
version_compare is all we need.

Change-Id: I33d4e31ba51ce9ee5e185fb5d15692c15fd7fcdc