[TASK] Allow unfinished appointments to be resumed
When an unfinished appointment is abandoned by moving to another page,
there was no supported way to resume it. This change should produce an
unfinished-list on the listpage of a user, if enabled by TS. It lists
all personally created appointments (not past ones) that have the
unfinished state set. Clicking their links allows you to resume the
creation process.

@TODO:
- Warning for lack of support, removed by script checking sessionStorage
support
- Folding of message-box
- Remove all button which marks the entire list as DELETED, with a
confirm box
- icon/message indicating incomplete data, per appointment, removed by
script checking sessionStorage data

The timer script should be changed to support multiple timers on a
single page. Also, it should differentiate between the two types of
timers, because they wont have the same visual effect when reaching 0. A
timer per appointment should be supported, even though caching prevents
good use for it, for now.

The SessionStorage should also work differently now:
- it can store form-data per appointment-id, so its id is now included
in the form-id. A consequence is that we can no longer clear the
SessionStorage, as doing so _selectively_ becomes extremely difficult
and a waste of resources. This poses a problem for the edit-form. But
this is already solved because..
- it is disabled for the edit-form. It was conceived for containing
formdata that is not yet persisted through the new-form, not while
editing an already persisted appointment. To differentiate, form class
.session now reflects whether the sessionStorage-mechanism is to be
initiated for that particular form, regardless if it contains .session
fields, while the latter still indicate which fields exactly are stored.

The ids and classes of forms have been changed to support the above
changes.

Change-Id: I7ed17b3a3d11b1cc5c1a5ae92d339d969e279793