[TASK] Add ignorevalidation annotations
Dontvalidate is broken on the deprecatedPropertyMapper since 6.2, when
using a model-specific validator. When searching for bugreports, I found
out about @ignorevalidation replacing @dontvalidate. This is not the
cause of the former though, but since the latter has come to our
attention, this needs to be settled at some point either way, so: add
them where you'd add @dontvalidate.

Change-Id: Iea917f5b1f995397fd42b673bd87faea9a7aa57e