[TASK] Make Textfield VHs 6.2 compatible
Between TYPO3 6.1 and 6.2, the placeholder attribute for the Textfield
VH was moved from a render-argument to one initialized in the
init-method. To be compatible with both versions, we need te rename the
placeholder argument and simply pass it along the parent for <6.2, while
registering its value as the placeholder tag-attribute for 6.2.

Change-Id: I77ce1916bf2cf90babae70f2a32a14b56ddfdbb1