[BUGFIX] TYPO3 6.2 form>for>checkbox int-property halts for
In TYPO3 6.2, an int subproperty name will result in a foreach on that
property name. If the property is also subject of a for VH encapsulating
the checkbox VH, this will set the array pointer on the objectStorage to
its end, prematurely ending the for VH.

A quick solution is to not let the subproperty names be ints.
Thankfully, I've only used those subproperties names to differentiate
between the items (so their name is irrelevant), so prefixing them with
a random character already solves the issue.

Change-Id: I3d0b1a38d57850ad1d8dc0cab2e5ca1d620a4b2b