[BUGFIX] CSRF security always fails in chrome/safari
The issue is that those browser send an ORIGIN header, for which I added
support. The header has the baseUrl without trailing slash however,
meaning a needle with trailing slash and a haystack without, will result
in a failed header validation.

By removing the trailing slash from the baseUrl in the header
validation, this is resolved and still compatible with the REFERER
header.

Change-Id: I6a8db789ca335d85aaff003004b6a7585eb24901