[BUGFIX] Private hash is not shared between sessions
The CSRF Protection logic already takes care of when to generate a new
private hash and when to get the last generated one. So why are we
storing the privateHash in the non-persisted sessiondata? (key 'ses' as
opposed to 'user')

The problem with 'ses' is that that both methods BASIC and STRONG_HEADER
fail once you either close your browser-session and try again within the
lifetime of the cache, or when you try it from a different
browser-session altogether.

This can be fixed by setting the key to 'user', so that it actually gets
persisted between seperate browser-sessions for the user.

Change-Id: Ib5625f9710df34db624b140542817ffe6fab9560