[TASK] Move and rename the CSRF constants
The constants are moved from the abstract to the interface.

Furthermore, the naming needs adjustments: STRONG_HEADER isn't a good
way to describe BASIC + REFERRER dependency, and MAXIMUM isn't a good
way to describe STRONG_JS + REFERRER dependency. Also, STRONG_UNCACHED
isn't of the same level as STRONG_JS. The new naming is:

Maximum+ (javascript / referrer)
Maximum (javascript)
Strong+ (no caching / referrer)
Strong (no caching)
Basic+ (referrer)
Basic
Disabled

The difference between Maximum and Strong is now the JavaScript layer.
The difference between Strong and Basic is token-per-request vs.
token-per-session respectively.
The REFERRER dependency is more of a bonus of added randomization and
requirement of an existing and matching header, so it is indicated as a
plus.

Change-Id: Ifb9357ac9486bb11c79ed2dacdfd7dcd0b5f9848