[TASK] Introduce timed hashes, less implementation in abstract
Hashes need to have a Time To Live (TTL), after which they expire. This
can easily be done by registering the time of generation, and
incorporating a isHashStillValid() method on retrieval from session.

Introducing timed hashes shows that we have still too much
implementation in the abstract as well, which cause more and more
redundant overrules. So a side-affect to this task is to move more
implementation-logic out of the abstract.

(NOTE to self: this could have been done better in two tasks)

Change-Id: I52788170c3c966cf59045cbf232c4d7065dfd706