[BUGFIX] CSRF protect on MAX+ fails
The hashsource is calculated with regards to the requestUri, which in
case of ajax calls is not the same as the ajax-calling page. But the
ajax call will have the same referer header as the expected requestUri,
so we can fix this issue by simply adding a check for JS-dependency on
where to retrieve the sourceUri for the hashSource.

Change-Id: Iede6952bd73b3df455e9998847178e60d0c6346f