[TASK] Make XHR pageID + Type configurable
While we haven't moved to eid yet, the xhr pagetype needs to be
configurable. This means a TS setting to influence the pageType.

Just as important is the pageId.. we can find everything we need via
window.location, but I'd rather take it to an eid script that removes
the need for a pageid. Until then, we can set the pageId with the TS
constant plugin.tx_appointments.settings.pages.xhr

Change-Id: Icad172f005f928096e13675550776f515d179021