[TASK] Split the 2-way Exclusive Availability in 2 settings
Types have a setting to be completely exclusive of other types. This
means they can be planned regardless of other types on the same agenda.

At some point, we made it 2-way, effectively making it invisible to
other types as well. This was to prevent the situation where a newly
planned exclusive type would prevent us from allowing to edit an earlier
existing non-exclusive type, when you've blocked its time.

These 2 characteristics need to be split, because main sponsor has the
desire to have the first without the latter in some cases.

Change-Id: I30bbf3da29d2a6139a912c4805e8baaea72fd574