[BUGFIX] Fix complex foreign_table_where
Queries with complex foreign_table_where would consistently fail in
TYPO3 8.7.4, because their ORDER BY or GROUP BY would not get extracted
by core before putting it in an enclosed WHERE.

This is because the regular expression employed for this use-case is too
simple: it does not expect newlines and possibly tabs, causing
non-matches where there should be matches.

The only real solution is to put the entire foreign_table_where into a
single line, but their complexity make them unreadable/unmaintainable to
the human eye this way.

So be it. I don't have time to go through patching the TYPO3 core before
fixing this extension.

Change-Id: I744353f06f20cac18a7ac4471577b32b325388cb