[TASK] Get rid of our own FeUser objects
These were added back in 4.5 because recordType was required to be set
to Extbase for anything using the Extbase domain models/repository, and
the configuration setting for this didn't do anything, at least up to
4.7.

Tests indicate we don't need them in T3v8. Any feuser record can be
mapped to an extbase domain model for FrontendUser.

Change-Id: I7acffe9d7c7bdf16dab1e26b2a3709437e0af533