[TASK] Use AssetProvider for JS assets
Includes jQuery UI CSS. We're not going to include the frontend CSS for
now, since that should be taken into concatenation/compression and a
single cachable version of that file is more efficient than multiple
versions. We ARE including the jQuery UI CSS because it isn't supposed
to be included with this anyway due to related files, so it would be a
separate style element linked from its original location.

Change-Id: I08d66c146e7a6083aac1e48af38608ac713730c8