[BUGFIX] Enablefields don't work
FormField.enableField is lazy. Somehow, this prevents TYPO3v9's Fluid to
access its uid property, as if the object never initializes in this
context once accessed. The issue does not exist if you explicitly debug,
or if you use the f:debug VH beforehand, as that will initialize the
lazy object prior to accessing the property. So maybe this is somehow a
PHP 7.3 bug, I can't tell as of yet.